+++ src/global.h
@@ -19,16 +19,16 @@ extern int newNNB;
 //extern int NumNeighborMax;
 
 // Time
-extern double global_time;
-extern double global_time_irr;
+extern REAL global_time;
+extern REAL global_time_irr;
 extern ULL NextRegTimeBlock;
 extern int time_block;
-extern double time_step;
+extern REAL time_step;
 extern ULL block_max;
 
 
-extern double binary_time;
-extern double binary_time_prev;
+extern REAL binary_time;
+extern REAL binary_time_prev;
 extern ULL binary_block;
 
 // ComputationChain
@@ -42,23 +42,23 @@ extern std::vector<Binary*> BinaryList; // List of binaries to calculate
 
 //extern bool debug;
 extern char* fname;
-extern double inputTime;
-extern double endTime;
+extern REAL inputTime;
+extern REAL endTime;
 extern bool restart;
 
 
 // Enzo to Nbody
 extern Particle* FirstEnzoParticle;
-extern double EnzoLength, EnzoMass, EnzoVelocity, EnzoTime, EnzoForce, EnzoAcceleration;
-extern double EnzoTime;
-extern double EnzoTimeStep;
+extern REAL EnzoLength, EnzoMass, EnzoVelocity, EnzoTime, EnzoForce, EnzoAcceleration;
+extern REAL EnzoTime;
+extern REAL EnzoTimeStep;
 
 
 // i/o
 extern char* foutput;
 extern bool IsOutput;
-extern double outputTime;
-extern double outputTimeStep;
+extern REAL outputTime;
+extern REAL outputTimeStep;
 extern int outNum;
 //
 //
